removed atomic<long double>
As it doesnt link on linux and is probably nonsense

+++ modules/napaudio/src/audio/node/bufferplayernode.h
@@ -85,7 +85,7 @@ namespace nap
   
             std::atomic<bool> mPlaying = { false }; // Indicates wether the node is currently playing.
             std::atomic<int> mChannel = { 0 }; // The channel within the buffer that is being played bacl/
-            std::atomic<long double> mPosition = { 0 }; // Current position of playback in samples within the source buffer.
+            std::atomic<double> mPosition = { 0 }; // Current position of playback in samples within the source buffer.
             std::atomic<ControllerValue> mSpeed = { 1.f }; // Playback speed as a fraction of the original speed.
             SafePtr<MultiSampleBuffer> mBuffer = nullptr; // Pointer to the buffer with audio material being played back.
         };
